packages/auth-oauth2/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "yagura-auth-oauth2"
version = "0.1.0"
description = "OAuth2/OIDC AuthProvider for Yagura (Google, Azure AD, Okta, Auth0, Keycloak)."
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.12"
authors = [
    { name = "Kouta", email = "sk-kouta-dev@users.noreply.github.com" }
]
license-files = ["LICENSE"]
keywords = ["ai", "agent", "yagura", "tools"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "yagura-agent>=0.1.0",
    "authlib>=1.3",
    "httpx>=0.25",
    "pyjwt[crypto]>=2.8",
]

[project.urls]
Homepage = "https://github.com/sk-kouta-dev/yagura"
Documentation = "https://github.com/sk-kouta-dev/yagura"
Repository = "https://github.com/sk-kouta-dev/yagura"

[tool.hatch.build.targets.wheel]
packages = ["yagura_auth"]
